Moving files into data subdirectory in account-settings
Including all the utils stuff.
