+++ src/account-settings/EmailField.jsx
@@ -13,8 +13,8 @@ import messages from './AccountSettingsPage.messages';
 import {
   openForm,
   closeForm,
-} from './actions';
-import { editableFieldSelector } from './selectors';
+} from './data/actions';
+import { editableFieldSelector } from './data/selectors';
 
 
 function EmailField(props) {
