+++ src/account-settings/data/reducers.js
@@ -9,10 +9,10 @@ import {
   RESET_DRAFTS,
 } from './actions';
 
-import { reducer as deleteAccountReducer, DELETE_ACCOUNT } from './delete-account';
-import { reducer as siteLanguageReducer, FETCH_SITE_LANGUAGES } from './site-language';
-import { reducer as resetPasswordReducer, RESET_PASSWORD } from './reset-password';
-import { reducer as thirdPartyAuthReducer, DISCONNECT_AUTH } from './third-party-auth';
+import { reducer as deleteAccountReducer, DELETE_ACCOUNT } from '../delete-account';
+import { reducer as siteLanguageReducer, FETCH_SITE_LANGUAGES } from '../site-language';
+import { reducer as resetPasswordReducer, RESET_PASSWORD } from '../reset-password';
+import { reducer as thirdPartyAuthReducer, DISCONNECT_AUTH } from '../third-party-auth';
 
 export const defaultState = {
   loading: false,
