+++ src/account-settings/delete-account/data/actions.js
@@ -1,4 +1,4 @@
-import { AsyncActionType } from '../../utils';
+import { AsyncActionType } from '../../data/utils';
 
 export const DELETE_ACCOUNT = new AsyncActionType('ACCOUNT_SETTINGS', 'DELETE_ACCOUNT');
 DELETE_ACCOUNT.CONFIRMATION = 'ACCOUNT_SETTINGS__DELETE_ACCOUNT__CONFIRMATION';
