+++ src/account-settings/index.js
@@ -1,5 +1,5 @@
 export { default } from './AccountSettingsPage';
-export { default as reducer } from './reducers';
-export { default as saga } from './sagas';
-export { storeName } from './selectors';
+export { default as reducer } from './data/reducers';
+export { default as saga } from './data/sagas';
+export { storeName } from './data/selectors';
 export { default as NotFoundPage } from './NotFoundPage';
