+++ src/account-settings/reset-password/data/actions.js
@@ -1,4 +1,4 @@
-import { AsyncActionType } from '../../utils';
+import { AsyncActionType } from '../../data/utils';
 
 export const RESET_PASSWORD = new AsyncActionType('ACCOUNT_SETTINGS', 'RESET_PASSWORD');
