+++ src/account-settings/site-language/selectors.js
@@ -1,5 +1,5 @@
 import { createSelector } from 'reselect';
-import { getModuleState } from '../utils';
+import { getModuleState } from '../data/utils';
 
 export const storePath = ['accountSettings', 'siteLanguage'];
